Move convertSnakeCaseToCamelCase to utils
